Removes NotificationActions component and consolidates logic
Deletes NotificationActions component and its tests, moving key functionalities to NotificationCard.
